+++ checkout/src/main/java/company/tap/checkout/viewholders/CardViewHolder.kt
@@ -44,7 +44,7 @@ class CardViewHolder(private val context: Context) : TapBaseViewHolder {
     fun setThemeToLeftButton(){
         val buttonTheme = TextViewTheme()
         buttonTheme.textColor = Color.parseColor(ThemeManager.getValue("horizontalList.headers.gatewayHeader.leftButton.labelTextColor"))
-        buttonTheme.textSize = ThemeManager.getFontSize("horizontalList.headers.gatewayHeader.leftButton.labelTextFont").toFloat()
+        buttonTheme.textSize = ThemeManager.getFontSize("horizontalList.headers.gatewayHeader.leftButton.labelTextFont")
         buttonTheme.font = ThemeManager.getFontName("horizontalList.headers.gatewayHeader.leftButton.labelTextFont")
         view.groupName.setTheme(buttonTheme)
     }
@@ -52,7 +52,7 @@ class CardViewHolder(private val context: Context) : TapBaseViewHolder {
     fun setThemeToRightButton(){
         val buttonTheme = TextViewTheme()
         buttonTheme.textColor = Color.parseColor(ThemeManager.getValue("horizontalList.headers.gatewayHeader.rightButton.labelTextColor"))
-        buttonTheme.textSize = ThemeManager.getFontSize("horizontalList.headers.gatewayHeader.rightButton.labelTextFont").toFloat()
+        buttonTheme.textSize = ThemeManager.getFontSize("horizontalList.headers.gatewayHeader.rightButton.labelTextFont")
         buttonTheme.font = ThemeManager.getFontName("horizontalList.headers.gatewayHeader.rightButton.labelTextFont")
         view.groupAction.setTheme(buttonTheme)
     }
