+++ checkout/src/main/java/company/tap/checkout/viewholders/SwitchViewHolder.kt
@@ -44,7 +44,7 @@ class SwitchViewHolder(context: Context) : TapBaseViewHolder {
     fun setThemeView(){
         view.switchSaveMobile.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
         var switchTheme = SwitchTheme()
-        switchTheme.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
+        switchTheme.textSize = ThemeManager.getFontSize("TapSwitchView")
         switchTheme.textColor = Color.parseColor(ThemeManager.getValue("TapSwitchView"))
         view.switchSaveMobile.setTheme(switchTheme)
 
@@ -52,21 +52,21 @@ class SwitchViewHolder(context: Context) : TapBaseViewHolder {
         view.saveTextView.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
          var textViewTheme = TextViewTheme()
         textViewTheme.textColor = Color.parseColor(ThemeManager.getValue("TapSwitchView"))
-        textViewTheme.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
+        textViewTheme.textSize = ThemeManager.getFontSize("TapSwitchView")
         textViewTheme.font = ThemeManager.getFontName("TapSwitchView")
         view.saveTextView.setTheme(textViewTheme)
 
 
         view.switchSaveMerchant.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
         var switchTheme1 = SwitchTheme()
-        switchTheme1.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
+        switchTheme1.textSize = ThemeManager.getFontSize("TapSwitchView")
         switchTheme1.textColor = Color.parseColor(ThemeManager.getValue("TapSwitchView"))
         view.switchSaveMerchant.setTheme(switchTheme1)
 
 
         view.switchgoPayCheckout.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
         var switchTheme2 = SwitchTheme()
-        switchTheme2.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
+        switchTheme2.textSize = ThemeManager.getFontSize("TapSwitchView")
         switchTheme2.textColor = Color.parseColor(ThemeManager.getValue("TapSwitchView"))
         view.switchgoPayCheckout.setTheme(switchTheme2)
 
@@ -74,7 +74,7 @@ class SwitchViewHolder(context: Context) : TapBaseViewHolder {
         view.savegoPay.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
         var textViewTheme1 = TextViewTheme()
         textViewTheme1.textColor = Color.parseColor(ThemeManager.getValue("TapSwitchView"))
-        textViewTheme1.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
+        textViewTheme1.textSize = ThemeManager.getFontSize("TapSwitchView")
         textViewTheme1.font = ThemeManager.getFontName("TapSwitchView")
         view.savegoPay.setTheme(textViewTheme1)
 
@@ -82,7 +82,7 @@ class SwitchViewHolder(context: Context) : TapBaseViewHolder {
         view.alertgoPaySignup.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
         var textViewTheme2 = TextViewTheme()
         textViewTheme2.textColor = Color.parseColor(ThemeManager.getValue("TapSwitchView"))
-        textViewTheme2.textSize = ThemeManager.getFontSize("TapSwitchView").toFloat()
+        textViewTheme2.textSize = ThemeManager.getFontSize("TapSwitchView")
         textViewTheme2.font = ThemeManager.getFontName("TapSwitchView")
         view.alertgoPaySignup.setTheme(textViewTheme2)
